Deleted old "info menu" class
This class had nothing on it, although it was used for a summary of the plot.
It consisted of:
"A silent bounty hunter named Peast is on a quest to kill a strong monster called "Freshiza". Peast will face many dangers on his way like snakes, soldiers called they "Zayne", spooky ghosts, time limits and much more. 
Fight them all to gain experience and utilize the leveling up system which can allow Peast to be more efficient and powerful fighter!
Can you rise up to the challenge?
Or will the challenge rise up to you?"